Remove BasePluginConvention and a few methods in BasePluginExtension
