+++ platforms/software/platform-base/src/main/java/org/gradle/api/plugins/BasePluginExtension.java
@@ -21,8 +21,6 @@ import org.gradle.api.provider.Property;
 
 /**
  * An extension used for {@link BasePlugin}.
- * <p>
- * Replaces {@link BasePluginConvention}.
  *
  * @since 7.1
  */
@@ -45,52 +43,4 @@ public interface BasePluginExtension {
      * The base name to use for archive files.
      */
     Property<String> getArchivesName();
-
-    /**
-     * This method is only here to maintain compatibility with existing builds.
-     *
-     * @deprecated Use {@link #getDistsDirectory()}. This method is scheduled for removal in Gradle 9.0.
-     */
-    @Deprecated
-    String getDistsDirName();
-
-    /**
-     * This method is only here to maintain compatibility with existing builds.
-     *
-     * @deprecated Use {@link #getDistsDirectory()}. This method is scheduled for removal in Gradle 9.0.
-     */
-    @Deprecated
-    void setDistsDirName(String distsDirName);
-
-    /**
-     * This method is only here to maintain compatibility with existing builds.
-     *
-     * @deprecated Use {@link #getLibsDirectory()}. This method is scheduled for removal in Gradle 9.0.
-     */
-    @Deprecated
-    String getLibsDirName();
-
-    /**
-     * This method is only here to maintain compatibility with existing builds.
-     *
-     * @deprecated Use {@link #getLibsDirectory()}. This method is scheduled for removal in Gradle 9.0.
-     */
-    @Deprecated
-    void setLibsDirName(String libsDirName);
-
-    /**
-     * This method is only here to maintain compatibility with existing builds.
-     *
-     * @deprecated Use {@link #getArchivesName()}. This method is scheduled for removal in Gradle 9.0.
-     */
-    @Deprecated
-    String getArchivesBaseName();
-
-    /**
-     * This method is only here to maintain compatibility with existing builds.
-     *
-     * @deprecated Use {@link #getArchivesName()}. This method is scheduled for removal in Gradle 9.0.
-     */
-    @Deprecated
-    void setArchivesBaseName(String archivesBaseName);
 }
