+++ platforms/software/platform-base/src/test/groovy/org/gradle/api/plugins/BasePluginTest.groovy
@@ -34,7 +34,6 @@ class BasePluginTest extends AbstractProjectBuilderSpec {
         project.pluginManager.apply(BasePlugin)
 
         then:
-        project.convention.plugins.base instanceof BasePluginConvention
         project.extensions.findByType(BasePluginExtension) != null
     }
 
@@ -142,15 +141,15 @@ class BasePluginTest extends AbstractProjectBuilderSpec {
 
         then:
         def someZip = project.tasks.create('someZip', Zip)
-        someZip.destinationDirectory.get().asFile == project.distsDirectory.get().asFile
+        someZip.destinationDirectory.get().asFile == project.base.distsDirectory.get().asFile
         someZip.archiveVersion.get() == project.version
-        someZip.archiveBaseName.get() == project.archivesBaseName
+        someZip.archiveBaseName.get() == project.base.archivesName.get()
 
         and:
         def someTar = project.tasks.create('someTar', Tar)
-        someTar.destinationDirectory.get().asFile == project.distsDirectory.get().asFile
+        someTar.destinationDirectory.get().asFile == project.base.distsDirectory.get().asFile
         someTar.archiveVersion.get() == project.version
-        someTar.archiveBaseName.get() == project.archivesBaseName
+        someTar.archiveBaseName.get() == project.base.archivesName.get()
     }
 
     def "uses null version when project version not specified"() {
