+++ platforms/software/platform-base/src/test/groovy/org/gradle/api/plugins/DefaultBasePluginExtensionTest.groovy
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2007 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.plugins
+
+import org.gradle.test.fixtures.AbstractProjectBuilderSpec
+
+class DefaultBasePluginExtensionTest extends AbstractProjectBuilderSpec {
+
+    private BasePluginExtension extension
+
+    def setup() {
+        project.pluginManager.apply(BasePlugin)
+        extension =  project.extensions.base
+    }
+
+    def "default values"() {
+        expect:
+        extension.distsDirectory.getAsFile().get() == project.layout.buildDirectory.dir('distributions').get().asFile
+    }
+}
