+++ subprojects/core/src/main/resources/org/gradle/initialization/converted-types.txt
@@ -1,7 +1,6 @@
 org.gradle.api.distribution.plugins.DistributionPlugin
 org.gradle.api.plugins.ApplicationPlugin
 org.gradle.api.plugins.ApplicationPluginConvention
-org.gradle.api.plugins.BasePluginConvention
 org.gradle.api.plugins.JavaLibraryDistributionPlugin
 org.gradle.api.plugins.JavaPluginConvention
 org.gradle.api.plugins.ProjectReportsPluginConvention
