+++ testing/architecture-test/src/changes/archunit-store/public-api-mutable-properties.txt
@@ -147,12 +147,6 @@ Method <org.gradle.api.plugins.ApplicationPluginConvention.getApplicationName()>
 Method <org.gradle.api.plugins.ApplicationPluginConvention.getExecutableDir()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (ApplicationPluginConvention.java:0)
 Method <org.gradle.api.plugins.ApplicationPluginConvention.getMainClassName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (ApplicationPluginConvention.java:0)
 Method <org.gradle.api.plugins.ApplicationPluginConvention.getProject()> does not have raw return type (org.gradle.api.Project) assignable to any of [Provider] in (ApplicationPluginConvention.java:0)
-Method <org.gradle.api.plugins.BasePluginConvention.getArchivesBaseName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (BasePluginConvention.java:0)
-Method <org.gradle.api.plugins.BasePluginConvention.getDistsDirName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (BasePluginConvention.java:0)
-Method <org.gradle.api.plugins.BasePluginConvention.getLibsDirName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (BasePluginConvention.java:0)
-Method <org.gradle.api.plugins.BasePluginExtension.getArchivesBaseName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (BasePluginExtension.java:0)
-Method <org.gradle.api.plugins.BasePluginExtension.getDistsDirName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (BasePluginExtension.java:0)
-Method <org.gradle.api.plugins.BasePluginExtension.getLibsDirName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (BasePluginExtension.java:0)
 Method <org.gradle.api.plugins.JavaApplication.getApplicationDefaultJvmArgs()> does not have raw return type (java.lang.Iterable) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (JavaApplication.java:0)
 Method <org.gradle.api.plugins.JavaApplication.getApplicationDistribution()> does not have raw return type (org.gradle.api.file.CopySpec) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (JavaApplication.java:0)
 Method <org.gradle.api.plugins.JavaApplication.getApplicationName()> does not have raw return type (java.lang.String) assignable to any of [Property, MapProperty, ListProperty, SetProperty] in (JavaApplication.java:0)
